Altered file structure. Scripts are now under server. cron job runs from server.js
